requirements.txt
pytest==7.4.4
selenium==4.39.0
pytest-selenium==4.1.0
allure-pytest==2.15.3
termcolor==3.2.0